requirements.txt
hydra-core==1.2.0
numpy==1.23.1
onnxruntime==1.12.1
pytorch-lightning==1.7.7